Add option to include public ssh key for first user
The script now has an option to include a public ssh key for the first
user to allow easy login via ssh with key authentication.

If a user is created, it will also set the password to the username,
force a password change on first login. The password change works even
if ssh key auth is done.