Add options for using LVM backingstore
These are just the basic options for picking lvm as the backingstore and
then optionally setting the volume group, logical volume name and volume
size. Without the other optional arguments, the lvm defaults will be
used.

At this stage the filesystem for lvm will default to ext4 which is fine
for now.